Move MAX_RESULT_RECORDS to constants.py for consistency
